Move client only info
