+++ src/universal/config/routes.ts
@@ -1,7 +1,4 @@
-import { generatePath } from 'react-router-dom';
-import { Match } from '../types';
-
-export const AppRoutes: Record<string, string> = {
+export const AppRoutes = {
   ROOT: '/',
   HOME: '/',
   BURGERZAKEN: '/burgerzaken',
@@ -71,69 +68,5 @@ export const AppRoutes: Record<string, string> = {
   ZAAK_STATUS: '/zaak-status',
 } as const;
 
-export const PublicRoutes: string[] = [
-  AppRoutes.API_LOGIN,
-  AppRoutes.API1_LOGIN,
-  AppRoutes.API2_LOGIN,
-  AppRoutes.ACCESSIBILITY,
-  AppRoutes.BFF_500_ERROR,
-  AppRoutes.ACCESSIBILITY,
-  AppRoutes.GENERAL_INFO,
-  AppRoutes.ROOT,
-];
-
-export const PrivateRoutes = Object.values(AppRoutes).filter(
-  (path: string) => !PublicRoutes.includes(path)
-);
-
 type RouteKey = keyof typeof AppRoutes;
 export type AppRoute = (typeof AppRoutes)[RouteKey];
-
-export interface TrackingConfig {
-  profileType: ProfileType;
-  isAuthenticated: boolean;
-}
-
-type CustomTrackingUrlMap = {
-  [key in AppRoute]+?: (match: Match, trackingConfig: TrackingConfig) => string;
-};
-
-export const CustomTrackingUrls: CustomTrackingUrlMap = {
-  [AppRoutes['VERGUNNINGEN/DETAIL']]: (match: Match) => {
-    return `/vergunning/${match.params?.title}`;
-  },
-
-  [AppRoutes['INKOMEN/BBZ']]: (match: Match) => {
-    return `/inkomen/bbz`;
-  },
-  [AppRoutes['INKOMEN/BIJSTANDSUITKERING']]: (match: Match) => {
-    return `/inkomen/bijstandsuitkering`;
-  },
-  [AppRoutes['INKOMEN/TOZO']]: (match: Match) => {
-    return `/inkomen/tozo/${match.params?.version}`;
-  },
-  [AppRoutes['INKOMEN/TONK']]: () => `/inkomen/tonk`,
-
-  [AppRoutes['BURGERZAKEN/ID-KAART']]: () => '/burgerzaken/id-kaart',
-
-  [AppRoutes['STADSPAS/AANVRAAG']]: () => '/stadspas/aavraag',
-  [AppRoutes['STADSPAS/SALDO']]: () => '/stadspas/saldo',
-
-  [AppRoutes['ZORG/VOORZIENINGEN']]: () => '/zorg-en-ondersteuning/voorziening',
-
-  [AppRoutes['TOERISTISCHE_VERHUUR/VERGUNNING/BB']]: () =>
-    '/toeristische-verhuur/vergunning/bed-and-breakfast',
-  [AppRoutes['TOERISTISCHE_VERHUUR/VERGUNNING/VV']]: () =>
-    '/toeristische-verhuur/vergunning/vakantieverhuur',
-  [AppRoutes['TOERISTISCHE_VERHUUR/VERGUNNING']]: () =>
-    '/toeristische-verhuur/vergunning',
-
-  [AppRoutes['KLACHTEN/KLACHT']]: () => '/klachten/klacht',
-
-  [AppRoutes['ERFPACHTv2/DOSSIERDETAIL']]: () => '/erfpacht/dossier',
-
-  [AppRoutes.ROOT]: (
-    match: Match,
-    { profileType, isAuthenticated }: TrackingConfig
-  ) => `/${isAuthenticated ? 'dashboard' : 'landing'}`,
-};
